# perf: bundle runtime dependencies