--- a/src/index.ts
+++ b/src/index.ts
@@ -5,6 +5,7 @@ export * from "./dev/server";
 export * from "./types";
 export * from "./prerender";
 export * from "./preset";
+export * from "./deps";
 
 export { loadOptions } from "./options";
 export type { LoadConfigOptions } from "./options";