--- a/src/types/nitro.ts
+++ b/src/types/nitro.ts
@@ -249,11 +249,11 @@ export interface NitroOptions extends PresetOptions {
     /**
      * Disable Experimental bundling of Nitro Runtime Dependencies
      */
-    inlineRuntimeDependencies?: false;
+    bundleRuntimeDependencies?: false;
     /**
      * Disable Experimental Sourcemap Minification
      */
-    sourcemapIgnore?: false;
+    sourcemapMinify?: false;
   };
   future: {
     nativeSWR: boolean;