<!---
☝️ PR title should follow conventional commits (https://conventionalcommits.org) 
-->

### 🔗 Linked issue

<!-- Please ensure there is an open issue and mention its number as #123 -->

### ❓ Type of change

<!-- What types of changes does your code introduce? Put an `x` in all the boxes that apply. -->

- [ ] 📖 Documentation (updates to the documentation, readme, or JSdoc annotations)
- [ ] 🐞 Bug fix (a non-breaking change that fixes an issue)
- [x] 👌 Enhancement (improving an existing functionality like performance)
- [ ] ✨ New feature (a non-breaking change that adds functionality)
- [ ] 🧹 Chore (updates to the build process or auxiliary tools and libraries)
- [ ] ⚠️ Breaking change (fix or feature that would cause existing functionality to change)

### 📚 Description

Nitro output uses file tracing for all node modules by default. It has many benefits such as supporting native modules and the best compatibility with minimized rollup involvement.

However, there are downsides with externals too at least for built-in and standard unjs packages that inlining is better.

- Issues with package hosting and multi-version handling
- we can optimize tree-shaking output up to file level instead of function exports 
- Extra data such as `package.json` files are copied over the output directory
- Extra startup overhead because of `node_modules` resolution
- Possible wrong export condition resolutions
- Inability to tree-shake with predictable platform injections (such as `AsyncLocalStorage`)

This PR uses a list of nitro built-in runtime dependencies and explicitly inlines them in the final output bundle.

TODO: This method however has a downside in that inlined sourcemaps add to the bundle overhead. Looking for a way to disable inline-sourcemaps only for node_modules .

### Benchmark Results

Measurement / Mode     | external deps | inlined deps
----------------|---------------|------------------
number of files | 120           | 13
disk size (without sourcemaps) | 840 KB         | 476 KB
start | 31.823 ms | 20ms
fetch | 14.18 ms | 14ms
start-to-fetch | 47.464ms | 35ms


Benchmark is not accurate and is made overnight!

Config:

```ts
export default defineNitroConfig({
  sourceMap: false,
  analyze: {
    filename: "./analyze.html",
  },
});
```

(stupid!) Timing bench:

```ts
console.time("start");
console.time("start-to-fetch");

await import("./b/server/index.mjs").then(async () => {
  console.timeEnd("start");
  console.time("fetch");
  await fetch("http://localhost:3000/").then((r) => {
    console.timeEnd("fetch");
    console.timeEnd("start-to-fetch");
  });
});
```

Node.js version: v18.16.0 on Macbook Air M2

### 📝 Checklist

<!-- Put an `x` in all the boxes that apply. -->
<!-- If your change requires a documentation PR, please link it appropriately -->
<!-- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] I have linked an issue or discussion.
- [ ] I have updated the documentation accordingly.