@pi0 2023-08-09
> In future we might need an opt out for major versions somehow? (If user, or other library is using dependency.) Or equally if one of these is a dependency of an external we might have an issue too.

Yes, i was thinking to support `externals.exclude` (and filter-out default inlines when user/framework specifically requires opt-out for a feature) + a flag to completely disable this behavior.

Regarding major versions (or any custom version by user), we still use rollup's search path, resolution, and aliases so overriding is possible by all means, however even totday, we require them (defined list) to be in Nitro-Expected semver-major as nitro internally uses all these deps. 

I want to make sure mainly it is good to you too for becoming the default behavior .