@danielroe 2023-08-09
It is fine to me as default behaviour. I'm observing the potential issue that if a library in externals uses ufo/pathe then this will be included twice (once in node_modules and once bundled). If other major versions of libraries like ufo are included in externals that entirely addresses my first point.