@pi0 2023-08-09
> if a library in externals uses ufo/pathe then this will be included twice (once in node_modules and once bundled).

That's a valid point if both inline and external use same semver/major of ufo. And was initial reason that we made all externals by default. I guess we can think of some ways to automatically "deoptimize" inlining however it involves traced file dependency "pre-analysis" and sub-dependencny deptimization.

I guess for default behavior, it should benefit most of users both Nitro and Nuxt with both bundle size but more importantly the stability of resolutions with rollup comparing to current externals.