@danielroe 2023-08-09
Yes, I see this as a good change, too.