@pi0 2023-08-15
Updated: Sourcemap size issue is (kinda) resolved by using a sourcemap-ignore plugin that scans emited sourcemap chunks and dynamically disables the ones that include node modules (reducing basic build from `2.1MB` to `454 kB`. 

I think there is way more area for improving this plugin. We can tree-shake content only for example. Considering we use lazy chunks for user code, this should be _relatively_ ok. This can be disabled with  `experimental.sourcemapIgnore = false` in case of any issues.